move HAM and PSO files to correct folders
